Refactor UI, rename components, and add DebugMenu
Regulated the UI elements and conducted a few changes. The UI components 'SearchComponent' and 'TabComponent' are renamed as 'SearchBar' and 'TabIcon' respectively. A DebugMenu was added to access the debug mode. The BaseMenu is moved to Common. Included all icons in a single asset. Revision updated in BetterChests.csproj. This update also accounted for pertinent amendments in the respective services and feature methods.