Updated menu context handling and restored locked item visibility
The code for handling different menu contexts in the ContainerFactory has been improved. It now handles `GameMenu` specifically by assigning the player as its context. Moreover, an update was made on the visibility of locked items within the main inventory menu, which was a feature that was previously removed. This change is particularly reflected in the LockItem file, where the condition for highlighting items has been tweaked.