Refactor search and UI code
Significant refactor in how the UI and search are handled in the code. Strain has been reduced on inventory functions by rearranging code and optimizing search functions. Furthermore, debugging has been improved with accurate and helpful hover texts. A small addendum was made in the localization for clearer user interaction.